+++ Packages/com.yontalane.dialogugui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.3] - 2025.08.07
+
+### Changed
+
+* Requires latest version of Dialog package.
+
 ## [1.0.2] - 2025.08.07
 
 ### Changed